#TODO: cross validation and grid search for deep svm architecture
This update doesn't read repeated data for classification (avoid the same file being end-classified with both classes), functions manually updated with better respect for the deep architecture of the dsvm. Decision functions are taken from the input layers, summed up and signed to be classified in the main layer. It's going to be great to have a cross-validation method that accepts more than one kernel.